requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.0
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.18.1